Addressed PR Changes. Need to Experimentally determine if Reflection is efficient
