+++ src/MongoDB.Analyzer/Core/Builders/AnalysisCodeGenerator.cs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using MongoDB.Analyzer.Core.Json;
+using MongoDB.Analyzer.Core.Utilities;
 using static MongoDB.Analyzer.Core.HelperResources.MqlGeneratorSyntaxElements.Builders;
 using static MongoDB.Analyzer.Core.HelperResources.ResourcesUtilities;
 
@@ -57,31 +59,7 @@ internal static class AnalysisCodeGenerator
             syntaxTrees.Add(s_renderer_2_19_and_higher);
         }
 
-        var compilation = CSharpCompilation.Create(
-            BuildersAnalysisConstants.AnalysisAssemblyName,
-            syntaxTrees,
-            referencesContainer.References,
-            new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary));
-
-        using var memoryStream = new MemoryStream();
-        var emitResult = compilation.Emit(memoryStream);
-
-        BuildersMqlGeneratorExecutor buildersMqlCodeExecutor = null;
-
-        if (emitResult.Success)
-        {
-            context.Logger.Log("Compilation successful");
-
-            memoryStream.Seek(0, SeekOrigin.Begin);
-
-            var mqlGeneratorType = DynamicTypeProvider.GetType(referencesContainer, memoryStream, MqlGeneratorFullName);
-
-            buildersMqlCodeExecutor = mqlGeneratorType != null ? new BuildersMqlGeneratorExecutor(mqlGeneratorType) : null;
-        }
-        else
-        {
-            context.Logger.Log($"Compilation failed with: {string.Join(Environment.NewLine, emitResult.Diagnostics)}");
-        }
+        var buildersMqlCodeExecutor = (BuildersMqlGeneratorExecutor)AnalysisCodeGeneratorUtilities.GetCodeExecutor(context, AnalysisType.Builders, syntaxTrees.ToArray());
 
         var result = new CompilationResult(
             buildersMqlCodeExecutor != null,
