+++ src/MongoDB.Analyzer/Core/Json/AnalysisCodeGenerator.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using static MongoDB.Analyzer.Core.HelperResources.JsonSyntaxElements.Json;
+using MongoDB.Analyzer.Core.Utilities;
 using static MongoDB.Analyzer.Core.HelperResources.ResourcesUtilities;
 
 namespace MongoDB.Analyzer.Core.Json;
@@ -23,16 +23,15 @@ internal static class AnalysisCodeGenerator
     private static readonly JsonGeneratorTemplateBuilder.SyntaxElements s_jsonGeneratorSyntaxElements;
     private static readonly ParseOptions s_parseOptions;
 
-
     static AnalysisCodeGenerator()
     {
-        s_helpersSyntaxTrees = GetCommonCodeResources();
+        s_helpersSyntaxTrees = GetCommonCodeResources(ResourceNames.Json.AttributeHandler);
         var jsonGeneratorSyntaxTree = GetCodeResource(ResourceNames.Json.JsonGenerator);
         s_jsonGeneratorSyntaxElements = JsonGeneratorTemplateBuilder.CreateSyntaxElements(jsonGeneratorSyntaxTree);
         s_parseOptions = jsonGeneratorSyntaxTree.Options;
     }
 
-    public static CompilationResult Compile(MongoAnalysisContext context, JsonExpressionAnalysis jsonExpressionAnalysis)
+    public static CompilationResult Compile(MongoAnalysisContext context, ExpressionsAnalysis jsonExpressionAnalysis)
     {
         var semanticModel = context.SemanticModelAnalysisContext.SemanticModel;
         var referencesContainer = ReferencesProvider.GetReferences(semanticModel.Compilation.References, context.Logger);
@@ -41,7 +40,7 @@ internal static class AnalysisCodeGenerator
             return CompilationResult.Failure;
         }
 
-        var typesSyntaxTree = TypesGeneratorHelper.GenerateTypesSyntaxTree(AnalysisType.Json, jsonExpressionAnalysis.TypesDeclarations, s_parseOptions);
+        var typesSyntaxTree = TypesGeneratorHelper.GenerateTypesSyntaxTree(AnalysisType.Poco, jsonExpressionAnalysis.TypesDeclarations, s_parseOptions);
         var jsonGeneratorSyntaxTree = GenerateJsonGeneratorSyntaxTree(jsonExpressionAnalysis);
 
         var syntaxTrees = new List<SyntaxTree>(s_helpersSyntaxTrees)
@@ -50,31 +49,7 @@ internal static class AnalysisCodeGenerator
                 jsonGeneratorSyntaxTree
             };
 
-        var compilation = CSharpCompilation.Create(
-            JsonAnalysisConstants.AnalysisAssemblyName,
-            syntaxTrees,
-            referencesContainer.References,
-            new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary));
-
-        using var memoryStream = new MemoryStream();
-        var emitResult = compilation.Emit(memoryStream);
-
-        JsonGeneratorExecutor jsonCodeExecutor = null;
-
-        if (emitResult.Success)
-        {
-            context.Logger.Log("Compilation successful");
-
-            memoryStream.Seek(0, SeekOrigin.Begin);
-
-            var jsonGeneratorType = DynamicTypeProvider.GetType(referencesContainer, memoryStream, JsonGeneratorFullName);
-
-            jsonCodeExecutor = jsonGeneratorType != null ? new JsonGeneratorExecutor(jsonGeneratorType) : null;
-        }
-        else
-        {
-            context.Logger.Log($"Compilation failed with: {string.Join(Environment.NewLine, emitResult.Diagnostics)}");
-        }
+        var jsonCodeExecutor = (JsonGeneratorExecutor)AnalysisCodeGeneratorUtilities.GetCodeExecutor(context, AnalysisType.Poco, syntaxTrees.ToArray());
 
         var result = new CompilationResult(
             jsonCodeExecutor != null,
@@ -84,14 +59,13 @@ internal static class AnalysisCodeGenerator
         return result;
     }
 
-    public static SyntaxTree GenerateJsonGeneratorSyntaxTree(JsonExpressionAnalysis jsonExpressionAnalysis)
+    public static SyntaxTree GenerateJsonGeneratorSyntaxTree(ExpressionsAnalysis jsonExpressionAnalysis)
     {
         var testCodeBuilder = new JsonGeneratorTemplateBuilder(s_jsonGeneratorSyntaxElements);
 
         foreach (var jsonContext in jsonExpressionAnalysis.AnalysisNodeContexts)
         {
-            var analysisNode = jsonContext.Node;
-            jsonContext.EvaluationMethodName = testCodeBuilder.AddPOCO(analysisNode.RewrittenPOCO);
+            jsonContext.EvaluationMethodName = testCodeBuilder.AddPoco(jsonContext.Node.RewrittenExpression as ClassDeclarationSyntax);
         }
 
         return testCodeBuilder.GenerateSyntaxTree();
