+++ src/MongoDB.Analyzer/Core/Linq/LinqAnalyzer.cs
@@ -129,7 +129,7 @@ internal static class LinqAnalyzer
             }
         }
 
-        return new AnalysisStats(mqlCount, internalExceptionsCount, driverExceptionsCount, compilationResult.MongoDBDriverVersion.ToString(3), null);
+        return new AnalysisStats(mqlCount, 0, internalExceptionsCount, driverExceptionsCount, compilationResult.MongoDBDriverVersion.ToString(3), null);
     }
 
     private static bool IsDriverOrLinqException(MQLResult mqlResult)
