+++ src/MongoDB.Analyzer/Core/Linq/LinqMqlGeneratorExecutor.cs
@@ -18,7 +18,7 @@ namespace MongoDB.Analyzer.Core.Linq;
 
 internal record MQLResult(string Mql, bool Linq3Only, Exception Exception);
 
-internal sealed class LinqMqlGeneratorExecutor
+internal sealed class LinqMqlGeneratorExecutor : GeneratorExecutor
 {
     private readonly Type _testClassType;
     private readonly bool _isLinq3Driver;
