+++ tests/MongoDB.Analyzer.Tests.Common.TestCases/Json/NotSupportedJsonClassDeclarations.cs
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using System;
+using System.Collections.Generic;
 using MongoDB.Analyzer.Tests.Common.DataModel;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
-using System;
-using System.Collections.Generic;
 
 namespace MongoDB.Analyzer.Tests.Common.TestCases.Json
 {
